+++ src/App.css
@@ -8,8 +8,16 @@ body {
   font-family: Impact, Charcoal;
   text-align: center;
   margin: 10px;
+  padding-top: 100px;
 }
 
 #back_home {
   display: none;
+}
+
+@media screen and (max-width: 480px) {
+  #title {
+    font-size: 50px;
+    padding-top: 50px;
+  }
 }
